% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_abbreviation.R
\name{f_abbreviation}
\alias{f_abbreviation}
\alias{ff_abbreviation}
\title{Abbreviate Strings}
\usage{
f_abbreviation(x, length = 5, ...)

ff_abbreviation(...)
}
\arguments{
\item{x}{A vector of text strings.}

\item{length}{The minimum length of the abbreviations.}

\item{\ldots}{Other arguments passed to \code{\link[base]{abbreviate}}.}
}
\value{
Returns a string vector with strings abbreviated.
}
\description{
A wrapper for \code{\link[base]{abbreviate}} for abbreviating strings.
}
\examples{
f_abbreviation(state.name)
f_abbreviation('Cool Variable')
}
\seealso{
\code{\link[base]{abbreviate}}
}
