% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_data.R
\name{f_data_abbreviation}
\alias{f_data_abbreviation}
\title{Convert Data (byte) Labels to an Abbreviated Form}
\usage{
f_data_abbreviation(x, ...)
}
\arguments{
\item{x}{A vector of data labels.  One of:
"Petabyte", "Exabyte", "Zettabyte", "Yottabyte") ignoring case or
retaining c("b", "B", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB")
with proper case.}

\item{\ldots}{ignored.}
}
\value{
Returns avector of abbreviated data units.
}
\description{
Convert a data label such as Gigabyte to an abbreviated form
like 'GB'.
}
\examples{
x <- c("Exabyte", "terabyte", "ZB", "PetaByte", "KB", "byte", "Gigabyte",
"Bit", "GB", "b")

f_data_abbreviation(x)
}
