% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_logical.R
\name{f_logical}
\alias{f_logical}
\alias{ff_logical}
\alias{f_response}
\alias{ff_response}
\title{True/False Convert Logical/Dummy Code}
\usage{
f_logical(x, true = "True", false = "False", ...)

ff_logical(...)

f_response(x, yes = "Yes", no = "No", ...)

ff_response(...)
}
\arguments{
\item{x}{A vector of logical or dummy integers.  This vector will be coerced
to logical.}

\item{true}{A value for \code{TRUE} elements.}

\item{false}{A value for  \code{FALSE} elements.}

\item{\ldots}{ignored.}
}
\value{
Returns a string of either "True"/"False" elements.
}
\description{
Coerce logical (\code{TRUE}, \code{FALSE}) or or dummy coded elements (0/1)
to "True"/"False" elements.  This function is most useful in plot scales.
}
\examples{
f_logical(c(TRUE, TRUE, FALSE))
f_logical(c(1, 1, 0, 1, 0, 0, NA))
f_logical(c(1, 0, 2, .3, -3))
f_logical(rnorm(20) > 0)
f_logical(rnorm(20) > 0, "A", "B")

\dontrun{
library(ggplot2)
library(dplyr)

## Without labels
data_frame(dummy = sample(c(TRUE, FALSE), 30, TRUE)) \%>\%
    count(dummy) \%>\%
    ggplot(aes(dummy, n)) +
        geom_bar(stat = 'identity')

## With labels
data_frame(dummy = sample(c(TRUE, FALSE), 30, TRUE)) \%>\%
    count(dummy) \%>\%
    ggplot(aes(dummy, n)) +
        geom_bar(stat = 'identity') +
        scale_x_discrete(labels = f_logical)
}
}
\seealso{
\code{\link[base]{prettyNum}}
}
