% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_ordinal.R
\name{f_ordinal}
\alias{f_ordinal}
\title{Add Ordinal Suffixes (-st, -nd, -rd, -th) to Numbers}
\usage{
f_ordinal(x)
}
\arguments{
\item{x}{A vector of numbers (or string equivalents).}
}
\value{
Returns a string vector with ordinal suffixes.
}
\description{
Add ordinal suffixes (-st, -nd, -rd, -th) to numbers.
}
\examples{
f_ordinal(1:10)
}

