% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_sign.R
\name{f_sign}
\alias{f_sign}
\alias{ff_sign}
\title{Format Numeric Signs}
\usage{
f_sign(x, positive = "+", negative = "-", zero = "", ...)

ff_sign(...)
}
\arguments{
\item{x}{A vector of values.}

\item{positive}{A string/value to insert in for positive values.}

\item{negative}{A string/value to insert in for negative values.}

\item{zero}{A string/value to insert in for zero values.}

\item{\ldots}{ignored.}
}
\value{
Returns a string of signs.
}
\description{
\code{f_sign} - Formats numeric values to just their sign
('-' == < 0, '+' == > 0, or '' == 0).
}
\examples{
f_sign(c(-10, 0, 10))
f_sign(c(-10, 0, 10), zero = 0)
## web based
f_sign(c(-10, 0, 10), '<b>+</b>', '<b>&ndash;</b>')
}
\seealso{
\code{\link[numform]{f_num}}
}
