% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_state.R
\name{f_state}
\alias{f_state}
\alias{ff_state}
\title{Format State Names as Abbreviations}
\usage{
f_state(x, ...)

ff_state(...)
}
\arguments{
\item{x}{A vector of states.}

\item{\ldots}{ignored.}
}
\value{
Returns a string of abbreviated states.
}
\description{
Formats a state name as the abbreviated form.
}
\examples{
f_state(c('Texas', 'New York', NA, 'New Jersey', 'Washington', 'Europe'))
}
