% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_text_bar.R
\name{f_text_bar}
\alias{f_text_bar}
\alias{ff_text_bar}
\title{Format Text Based Bar Plots}
\usage{
f_text_bar(x, symbol = "_", width = 20, ...)

ff_text_bar(...)
}
\arguments{
\item{x}{A numeric vector.}

\item{symbol}{A sumbol to use for the bars.}

\item{width}{The max width of the bar.}

\item{\ldots}{ignored.}
}
\value{
Returns a vector of concatenated symbols as a string that represent x% of the bar.
}
\description{
Use a text symbol to create scaled horizontal bar plots of numeric vectors.
}
\examples{
library(dplyr)

mtcars \%>\%
    count(cyl, gear) \%>\%
    group_by(cyl) \%>\%
    mutate(
        p = numform::f_pp(n/sum(n))
    ) \%>\%
    ungroup() \%>\%
    mutate(
        cyl = numform::fv_runs(cyl),
        ` ` = text_bar(n)  ## Overall
    ) \%>\%
    as.data.frame()

mtcars \%>\%
    count(cyl, gear) \%>\%
    group_by(cyl) \%>\%
    mutate(
        p = numform::f_pp(n/sum(n)),
        ` ` = text_bar(n) ## within groups
    ) \%>\%
    ungroup() \%>\%
    mutate(
        cyl = numform::fv_runs(cyl),
        ` ` = text_bar(n)
    ) \%>\%
    as.data.frame()


mtcars \%>\%
    count(cyl, gear) \%>\%
    group_by(cyl) \%>\%
    mutate(
        p = numform::f_pp(n/sum(n)),
        `within groups` = text_bar(n, width = 8, symbol = '#')
    ) \%>\%
    ungroup() \%>\%
    mutate(
        cyl = numform::fv_runs(cyl),
        `overall` = text_bar(n, width = 30, symbol = '-')
    ) \%>\%
    as.data.frame()

## Drop the ehaders
mtcars \%>\%
    count(cyl, gear) \%>\%
    group_by(cyl) \%>\%
    mutate(
        p = numform::f_pp(n/sum(n)),
        `   ` = text_bar(n, symbol = '#')
    ) \%>\%
    ungroup() \%>\%
    mutate(
        cyl = numform::fv_runs(cyl),
        ` ` = text_bar(n, symbol = '-')
    ) \%>\%
    as.data.frame()
}
