% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_bills.R
\name{f_bills}
\alias{f_bills}
\alias{f_mills}
\alias{f_thous}
\title{Abbreviate Numbers}
\usage{
f_bills(x, relative = 0, digits = -9)

f_mills(x, relative = 0, digits = -6)

f_thous(x, relative = 0, digits = -3)
}
\arguments{
\item{x}{A vector of large numbers.}

\item{relative}{A factor relative to the current \code{digits} being rounded.
For example \code{relative = -1} moves one to the left while
\code{relative = 1} moves one to the right.}

\item{digits}{The number of digits to round to.  Actual \code{digits}
calculated as \code{digits} +  \code{relative}.}
}
\value{
Returns an abbreviated vector of numbers.
}
\description{
Use the K (thousands), M (millions), and B (billions) with abbreviated numbers.
}
\examples{
f_thous(1234)
f_thous(12345)
f_thous(123456)
f_mills(1234567)
f_mills(12345678)
f_mills(123456789)
f_bills(1234567891)
f_bills(12345678912)
f_bills(123456789123)

f_bills(123456789123, -1) # round to tens
f_bills(123456789123, -2) # round to hundreds
f_bills(123456789123, +1) # round to tenths
f_bills(123456789123, +2) # round to hundreths
\dontrun{
f_bills(123456789123, -2) \%>\%
    f_prefix("$")
}
}

