% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numform-package.R
\docType{package}
\name{numform}
\alias{numform}
\alias{package-numform}
\alias{numform-package}
\title{Tools to Format Numbers for Publication}
\description{
Format numbers for publication; includes the removal of leading zeros,
standardization of number of digits, and a p-value formatter.  These tools
combine the functionality of several 'base' functions such as \code{paste},
\code{format}, and \code{sprintf} into specific functions that are named in
a way that is consistent with usage, making them easy to remember and use.
}
