% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constant_months}
\alias{constant_months}
\alias{constant_months_abbreviation}
\alias{constant_weekdays}
\alias{constant_quarters}
\title{Constants}
\format{An object of class \code{character} of length 12.}
\usage{
constant_months

constant_months_abbreviation

constant_weekdays

constant_quarters
}
\description{
\code{constant_months}A constant for ordered month names.

\code{constant_months_abbreviation} - A constant for ordered month abbreviations.

\code{constant_weekdays} - A constant for ordered weekdays.

\code{constant_quarters} - A constant for ordered quarters.
}
\keyword{datasets}
