% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_month.R
\name{f_month}
\alias{f_month}
\alias{f_number2month}
\title{Format Months to One Letter Abbreviation}
\usage{
f_month(x)

f_number2month(x)
}
\arguments{
\item{x}{A vector of month names or integers 1-12.}
}
\value{
Returns a single letter month abbreviation atomic vector.
}
\description{
Format long month name or integer formats to a single capital letter.  Useful
for plot scales as a way to save space.  \code{f_month} is used for
nominal data where as  \code{f_number2month} is reserved for integers
1-12.
}
\examples{
f_month(month.name)
f_number2month(1:12)
}

