% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_sign.R
\name{f_sign}
\alias{f_sign}
\title{Format Numeric Signs}
\usage{
f_sign(x, ...)
}
\arguments{
\item{x}{A vector of values.}

\item{\ldots}{ignored.}
}
\value{
Returns a string of signs.
}
\description{
\code{f_sign} - Formats numeric values to just their sign
('-' == < 0, '+' == > 0, or '' == 0).
}
\examples{
f_sign(c(-10, 0, 10))
}
\seealso{
\code{\link[numform]{f_num}}
}

