% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv_percent.R
\name{fv_percent}
\alias{fv_percent}
\title{Convert a Numeric Vector to Percentages}
\usage{
fv_percent(x, digits = getOption("numformdigits"), ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{The number of digits to use.  Defaults to 1.  Can be set
globally via: \code{options(numformdigits = n)} where n is the number of
digits beyond the decimal point to include.}

\item{\ldots}{Other arguments passed to \code{\link[numform]{f_prop2percent}}.}
}
\value{
Returns a string of publication ready relative percentages.
}
\description{
Converts a numeric vector into a vector of relative percentages.
}
\examples{
fv_percent(1:4)
fv_percent(sample(1:100, 20))
}

