% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_bills.R
\name{f_bills}
\alias{f_bills}
\alias{ff_bills}
\alias{f_mills}
\alias{ff_mills}
\alias{f_thous}
\alias{ff_thous}
\title{Abbreviate Numbers}
\usage{
f_bills(x, relative = 0, digits = -9, prefix = "", ...)

ff_bills(...)

f_mills(x, relative = 0, digits = -6, prefix = "", ...)

ff_mills(...)

f_thous(x, relative = 0, digits = -3, prefix = "", ...)

ff_thous(...)
}
\arguments{
\item{x}{A vector of large numbers.}

\item{relative}{A factor relative to the current \code{digits} being rounded.
For example \code{relative = -1} moves one to the left while
\code{relative = 1} moves one to the right.}

\item{digits}{The number of digits to round to.  Actual \code{digits}
calculated as \code{digits} +  \code{relative}.}

\item{prefix}{A string to append to the front of elements.}

\item{\ldots}{ignored.}
}
\value{
Returns an abbreviated vector of numbers.
}
\description{
Use the K (thousands), M (millions), and B (billions) with abbreviated numbers.
}
\examples{
f_thous(1234)
f_thous(12345)
f_thous(123456)
f_mills(1234567)
f_mills(12345678)
f_mills(123456789)
f_bills(1234567891)
f_bills(12345678912)
f_bills(123456789123)

f_bills(123456789123, -1) # round to tens
f_bills(123456789123, -2) # round to hundreds
f_bills(123456789123, +1) # round to tenths
f_bills(123456789123, +2) # round to hundreths
\dontrun{
if (!require("pacman")) install.packages("pacman")
pacman::p_load(tidyverse, magrittr)

f_bills(123456789123, -2) \%>\%
    f_prefix("$")


data_frame(
    revenue = rnorm(100, 500000, 50000),
    deals = sample(20:50, 100, TRUE)
) \%>\%
    mutate(
        dollar = f_dollar(revenue, digits = -3),
        thous = f_thous(revenue),
        thous_dollars = f_thous(revenue, prefix = '$')
    ) \%T>\%
    print() \%>\%
    ggplot(aes(deals, revenue)) +
        geom_point() +
        geom_smooth() +
        scale_y_continuous(label = ff_thous(prefix = '$') )

data_frame(
    revenue = rnorm(10000, 500000, 50000),
    date = sample(seq(as.Date('1999/01/01'), as.Date('2000/01/01'), by="day"), 10000, TRUE),
    site = sample(paste("Site", 1:5), 10000, TRUE)
) \%>\%
    mutate(
        dollar = f_dollar(revenue, digits = -3),
        thous = f_thous(revenue),
        thous_dollars = f_thous(revenue, prefix = '$'),
        abb_month = f_month(date),
        abb_week = factor(f_weekday(date, distinct = TRUE),
        levels = c('Su', 'M', 'T', 'W', 'Th', 'F', 'S'))
    ) \%T>\%
    print() \%>\%
    ggplot(aes(abb_week, revenue)) +
        geom_jitter(width = .2, height = 0, alpha = .2) +
        scale_y_continuous(label = ff_thous(prefix = '$'))+
        facet_wrap(~site)
}
}
