% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numform-package.R
\docType{package}
\name{numform}
\alias{numform}
\alias{numform-package}
\alias{package-numform}
\title{Tools to Format Numbers for Publication}
\description{
Format numbers for publication; includes the removal of leading zeros, standardization of number of digits, and a p-value formatter.
}

