% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numform-package.R
\docType{package}
\name{numform}
\alias{numform}
\alias{package-numform}
\alias{numform-package}
\title{Tools to Format Numbers for Publication}
\description{
Format numbers and plots for publication; includes the removal of leading zeros,
standardization of number of digits, addition of affixes, and a p-value
formatter.  These tools combine the functionality of several 'base' functions
such as \code{\link[base]{paste}}, \code{\link[base]{format}}, and
\code{\link[base]{sprintf}} into specific use case functions that are named in
a way that is consistent with usage, making their names easy to remember and
easy to deploy.
}
